nm: Show bitrates with SI prefixes
I want to see the bitrate of my Ethernet connection as 1Gb/s or
100Mb/s rather than as e.g. 953.674316Mib/s. Similar for WiFi, there
are bitrates like 144.4Mb/s or 54Mb/s. With binary prefix, the former
is shown as 137.710571Mib/s.

Also set bitrate precision for WiFi to 0. The reason is that for some
bitrates we get a lot of trailing zeros. For example 144.4000000Mb/s.

Signed-off-by: Michal Sojka <michal.sojka@cvut.cz>