mem: Calculate better estimate of used memory
Mem plugin does not estimate the amount of used memory well. It is not
uncommon when the plugin shows memory utilization of 90%, while other
tools give numbers around 20%. The reason for such a big difference is
that the plugin uses sysinfo() system call, which does not provide all
information needed to calculate memory utilization precisely. Linux's
/proc/meminfo provides much better information on its MemAvailable
line (added in Linux 3.14). See the following commit for more
information:
https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=34e431b0ae398fc54ea69ff85ec700722c9da773

This changes the mem plugin to read the information from /proc/meminfo
and fall back to the sysinfo() method if /proc/meminfo is not
available or it does not contain MemAvailable line.

Now the values reported by the plugin are the same as values shown by
gnome-system-monitor tool.

Signed-off-by: Michal Sojka <michal.sojka@cvut.cz>