nm: Always refresh SSID after suspend/resume cycle
When suspending a laptop and resuming it at place with different Wi-Fi
AP, the SSID was not updated, i.e. the SSID of where the laptop was
suspended was shown all the time even when Wi-Fi connected to another
AP after resume.

This was caused by not updating section->ap to the new AP after
resuming. It was updated only after AP has changed while j4status was
running.

The problem could be reproduced this way:

1) Have two active APs that network manager (NM) can connect to.
2) Manually connect to the first AP.
3) Send j4status SIGUSR2 to stop it.
4) Manually connect to second AP.
5) Send j4status SIGUSR1 to let it continue (start).
6) Observe that j4status prints SSID of the first AP instead of the
   second one.

This is fixed by always setting section->ap in
_j4status_nm_device_monitor(), which is called from start
callback (after resume).

We also fix some reference counting problems.