cpu, fsinfo, iw: fix header checks
Apparently I missed the point of AC_SEARCH_LIBS when I first wrote those
plugins: it should, of course, raise an error if the headers aren't
found, and wrapping the include directives in conditionals would only
make error messages more cryptic.  Also, libblkid (for fsinfo) is a
proper library and is now linked.  Not sure how it worked before.