fsinfo, iw: more autoconf macros tweaks
This commit partially reverts the previous one: I had a bit of a brain
fart and hadn't noticed that libblkid was already linked through
PKG_CHECK_MODULES.  Really, the only reason I used the ugly
AC_SEARCH_LIBS macro in iw is because libiw lacks a pkg-config
interface.  And while I was there, I also changed that macro to a
AC_CHECK_LIB which I feel is more suited to this case.